feat: Enable KeePassXC open mode on Windows
