--- internal/cmd/keepassxctemplatefuncs_windows.go
@@ -1,7 +1,0 @@
-package cmd
-
-import "errors"
-
-func (c *Config) keepassxcOutputOpen(command string, args ...string) ([]byte, error) {
-	return nil, errors.New("not available on Windows")
-}
